Convert Config API to Stable endpoints.
