+++ examples/src/main/java/io/dapr/examples/configuration/http/ConfigurationClient.java
@@ -14,6 +14,7 @@ limitations under the License.
 package io.dapr.examples.configuration.http;
 
 import io.dapr.client.DaprApiProtocol;
+import io.dapr.client.DaprClient;
 import io.dapr.client.DaprClientBuilder;
 import io.dapr.client.DaprPreviewClient;
 import io.dapr.client.domain.ConfigurationItem;
@@ -42,8 +43,8 @@ public class ConfigurationClient {
    */
   public static void main(String[] args) throws Exception {
     System.getProperties().setProperty(Properties.API_PROTOCOL.getName(), DaprApiProtocol.HTTP.name());
-    try (DaprPreviewClient client = (new DaprClientBuilder()).buildPreviewClient()) {
-      System.out.println("Using preview client...");
+    try (DaprClient client = (new DaprClientBuilder()).build()) {
+      System.out.println("Using client...");
       getConfigurations(client);
       subscribeConfigurationRequest(client);
     }
@@ -54,7 +55,7 @@ public class ConfigurationClient {
    *
    * @param client DaprPreviewClient object
    */
-  public static void getConfigurations(DaprPreviewClient client) {
+  public static void getConfigurations(DaprClient client) {
     System.out.println("*******trying to retrieve configurations for a list of keys********");
     List<String> keys = new ArrayList<>();
     keys.add("myconfig1");
@@ -79,7 +80,7 @@ public class ConfigurationClient {
    *
    * @param client DaprPreviewClient object
    */
-  public static void subscribeConfigurationRequest(DaprPreviewClient client) throws InterruptedException {
+  public static void subscribeConfigurationRequest(DaprClient client) throws InterruptedException {
     System.out.println("Subscribing to key: myconfig2");
     SubscribeConfigurationRequest req = new SubscribeConfigurationRequest(
         CONFIG_STORE_NAME, Collections.singletonList("myconfig2"));
