+++ sdk-tests/src/test/java/io/dapr/it/configuration/http/ConfigurationIT.java
@@ -1,7 +1,7 @@
 package io.dapr.it.configuration.http;
 
+import io.dapr.client.DaprClient;
 import io.dapr.client.DaprClientBuilder;
-import io.dapr.client.DaprPreviewClient;
 import io.dapr.client.domain.ConfigurationItem;
 import io.dapr.it.BaseIT;
 import io.dapr.it.DaprRun;
@@ -21,7 +21,7 @@ public class ConfigurationIT extends BaseIT {
 
     private static DaprRun daprRun;
 
-    private static DaprPreviewClient daprPreviewClient;
+    private static DaprClient daprClient;
 
     private static String key = "myconfig1";
 
@@ -39,12 +39,12 @@ public class ConfigurationIT extends BaseIT {
     public static void init() throws Exception {
         daprRun = startDaprApp(ConfigurationIT.class.getSimpleName(), 5000);
         daprRun.switchToHTTP();
-        daprPreviewClient = new DaprClientBuilder().buildPreviewClient();
+        daprClient = new DaprClientBuilder().build();
     }
 
     @AfterClass
     public static void tearDown() throws Exception {
-        daprPreviewClient.close();
+        daprClient.close();
     }
 
     @Before
@@ -54,14 +54,14 @@ public class ConfigurationIT extends BaseIT {
 
     @Test
     public void getConfiguration() {
-        ConfigurationItem ci = daprPreviewClient.getConfiguration(CONFIG_STORE_NAME, "myconfigkey1").block();
+        ConfigurationItem ci = daprClient.getConfiguration(CONFIG_STORE_NAME, "myconfigkey1").block();
         assertEquals(ci.getKey(), "myconfigkey1");
         assertEquals(ci.getValue(), "myconfigvalue1");
     }
 
     @Test
     public void getConfigurations() {
-        Map<String, ConfigurationItem> cis = daprPreviewClient.getConfiguration(CONFIG_STORE_NAME, "myconfigkey1", "myconfigkey2").block();
+        Map<String, ConfigurationItem> cis = daprClient.getConfiguration(CONFIG_STORE_NAME, "myconfigkey1", "myconfigkey2").block();
         assertTrue(cis.size() == 2);
         assertTrue(cis.containsKey("myconfigkey1"));
         assertTrue(cis.containsKey("myconfigkey2"));
