+++ sdk/src/test/java/io/dapr/client/DaprPreviewClientGrpcTest.java
@@ -20,17 +20,12 @@ import com.google.protobuf.ByteString;
 import io.dapr.client.domain.BulkPublishRequest;
 import io.dapr.client.domain.BulkPublishEntry;
 import io.dapr.client.domain.BulkPublishResponse;
-import io.dapr.client.domain.ConfigurationItem;
 import io.dapr.client.domain.QueryStateItem;
 import io.dapr.client.domain.QueryStateRequest;
 import io.dapr.client.domain.QueryStateResponse;
-import io.dapr.client.domain.SubscribeConfigurationResponse;
-import io.dapr.client.domain.UnsubscribeConfigurationRequest;
-import io.dapr.client.domain.UnsubscribeConfigurationResponse;
 import io.dapr.client.domain.query.Query;
 import io.dapr.serializer.DaprObjectSerializer;
 import io.dapr.serializer.DefaultObjectSerializer;
-import io.dapr.v1.CommonProtos;
 import io.dapr.v1.DaprGrpc;
 import io.dapr.v1.DaprProtos;
 import io.grpc.Status;
@@ -47,19 +42,15 @@ import reactor.core.publisher.Mono;
 import java.io.Closeable;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
 
 import static io.dapr.utils.TestUtils.assertThrowsDaprException;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doAnswer;
@@ -72,7 +63,6 @@ import static org.mockito.Mockito.when;
 public class DaprPreviewClientGrpcTest {
 
 	private static final ObjectMapper MAPPER = new ObjectMapper();
-	private static final String CONFIG_STORE_NAME = "MyConfigStore";
 	private static final String QUERY_STORE_NAME = "testQueryStore";
 
 	private static final String PUBSUB_NAME = "testPubsub";
@@ -293,248 +283,6 @@ public class DaprPreviewClientGrpcTest {
 		BulkPublishResponse<String> result = previewClient.publishEvents(req).block();
 		Assert.assertNotNull(result);
 		Assert.assertEquals("expected no entries to be failed", 0, result.getFailedEntries().size());
-	}
-
-	@Test
-	public void getConfigurationTestErrorScenario() {
-		assertThrows(IllegalArgumentException.class, () -> {
-			previewClient.getConfiguration("", "key").block();
-		});
-	}
-
-	@Test
-	public void getSingleConfigurationTest() {
-		doAnswer((Answer<Void>) invocation -> {
-			StreamObserver<DaprProtos.GetConfigurationResponse> observer =
-					(StreamObserver<DaprProtos.GetConfigurationResponse>) invocation.getArguments()[1];
-			observer.onNext(getSingleMockResponse());
-			observer.onCompleted();
-			return null;
-		}).when(daprStub).getConfigurationAlpha1(any(DaprProtos.GetConfigurationRequest.class), any());
-
-		ConfigurationItem ci = previewClient.getConfiguration(CONFIG_STORE_NAME, "configkey1").block();
-		assertEquals("configvalue1", ci.getValue());
-		assertEquals("1", ci.getVersion());
-	}
-
-	@Test
-	public void getSingleConfigurationWithMetadataTest() {
-		doAnswer((Answer<Void>) invocation -> {
-			StreamObserver<DaprProtos.GetConfigurationResponse> observer =
-					(StreamObserver<DaprProtos.GetConfigurationResponse>) invocation.getArguments()[1];
-			observer.onNext(getSingleMockResponse());
-			observer.onCompleted();
-			return null;
-		}).when(daprStub).getConfigurationAlpha1(any(DaprProtos.GetConfigurationRequest.class), any());
-
-		Map<String, String> reqMetadata = new HashMap<>();
-		reqMetadata.put("meta1", "value1");
-		ConfigurationItem ci = previewClient.getConfiguration(CONFIG_STORE_NAME, "configkey1", reqMetadata).block();
-		assertEquals("configvalue1", ci.getValue());
-		assertEquals("1", ci.getVersion());
-	}
-
-	@Test
-	public void getMultipleConfigurationTest() {
-		doAnswer((Answer<Void>) invocation -> {
-			StreamObserver<DaprProtos.GetConfigurationResponse> observer =
-					(StreamObserver<DaprProtos.GetConfigurationResponse>) invocation.getArguments()[1];
-			observer.onNext(getMultipleMockResponse());
-			observer.onCompleted();
-			return null;
-		}).when(daprStub).getConfigurationAlpha1(any(DaprProtos.GetConfigurationRequest.class), any());
-
-		Map<String, ConfigurationItem> cis = previewClient.getConfiguration(CONFIG_STORE_NAME, "configkey1","configkey2").block();
-		assertEquals(2, cis.size());
-		assertTrue("configkey1", cis.containsKey("configkey1"));
-		assertEquals("configvalue1", cis.get("configkey1").getValue());
-		assertEquals("1", cis.get("configkey1").getVersion());
-		assertTrue("configkey2", cis.containsKey("configkey2"));
-		assertEquals("configvalue2", cis.get("configkey2").getValue());
-		assertEquals("1", cis.get("configkey2").getVersion());
-	}
-
-	@Test
-	public void getMultipleConfigurationWithMetadataTest() {
-		doAnswer((Answer<Void>) invocation -> {
-			StreamObserver<DaprProtos.GetConfigurationResponse> observer =
-					(StreamObserver<DaprProtos.GetConfigurationResponse>) invocation.getArguments()[1];
-			observer.onNext(getMultipleMockResponse());
-			observer.onCompleted();
-			return null;
-		}).when(daprStub).getConfigurationAlpha1(any(DaprProtos.GetConfigurationRequest.class), any());
-
-		Map<String, String> reqMetadata = new HashMap<>();
-		reqMetadata.put("meta1", "value1");
-		List<String> keys = Arrays.asList("configkey1","configkey2");
-		Map<String, ConfigurationItem> cis = previewClient.getConfiguration(CONFIG_STORE_NAME, keys, reqMetadata).block();
-		assertEquals(2, cis.size());
-		assertTrue("configkey1", cis.containsKey("configkey1"));
-		assertEquals("configvalue1", cis.get("configkey1").getValue());
-	}
-
-	@Test
-	public void subscribeConfigurationTest() {
-		Map<String, String> metadata = new HashMap<>();
-		metadata.put("meta1", "value1");
-		Map<String, CommonProtos.ConfigurationItem> configs = new HashMap<>();
-		configs.put("configkey1", CommonProtos.ConfigurationItem.newBuilder()
-		.setValue("configvalue1")
-		.setVersion("1")
-		.putAllMetadata(metadata)
-		.build());
-		DaprProtos.SubscribeConfigurationResponse responseEnvelope = DaprProtos.SubscribeConfigurationResponse.newBuilder()
-				.putAllItems(configs)
-				.setId("subscription_id")
-				.build();
-
-		doAnswer((Answer<Void>) invocation -> {
-			StreamObserver<DaprProtos.SubscribeConfigurationResponse> observer =
-					(StreamObserver<DaprProtos.SubscribeConfigurationResponse>) invocation.getArguments()[1];
-			observer.onNext(responseEnvelope);
-			observer.onCompleted();
-			return null;
-		}).when(daprStub).subscribeConfigurationAlpha1(any(DaprProtos.SubscribeConfigurationRequest.class), any());
-
-		Iterator<SubscribeConfigurationResponse> itr = previewClient.subscribeConfiguration(CONFIG_STORE_NAME, "configkey1").toIterable().iterator();
-		assertTrue(itr.hasNext());
-		SubscribeConfigurationResponse res = itr.next();
-		assertTrue(res.getItems().containsKey("configkey1"));
-		assertEquals("subscription_id", res.getSubscriptionId());
-		assertFalse(itr.hasNext());
-	}
-
-	@Test
-	public void subscribeConfigurationTestWithMetadata() {
-		Map<String, String> metadata = new HashMap<>();
-		metadata.put("meta1", "value1");
-		Map<String, CommonProtos.ConfigurationItem> configs = new HashMap<>();
-		configs.put("configkey1", CommonProtos.ConfigurationItem.newBuilder()
-		.setValue("configvalue1")
-		.setVersion("1")
-		.putAllMetadata(metadata)
-		.build());
-		DaprProtos.SubscribeConfigurationResponse responseEnvelope = DaprProtos.SubscribeConfigurationResponse.newBuilder()
-				.putAllItems(configs)
-				.setId("subscription_id")
-				.build();
-
-		doAnswer((Answer<Void>) invocation -> {
-			StreamObserver<DaprProtos.SubscribeConfigurationResponse> observer =
-					(StreamObserver<DaprProtos.SubscribeConfigurationResponse>) invocation.getArguments()[1];
-			observer.onNext(responseEnvelope);
-			observer.onCompleted();
-			return null;
-		}).when(daprStub).subscribeConfigurationAlpha1(any(DaprProtos.SubscribeConfigurationRequest.class), any());
-
-		Map<String, String> reqMetadata = new HashMap<>();
-		List<String> keys = Arrays.asList("configkey1");
-
-		Iterator<SubscribeConfigurationResponse> itr = previewClient.subscribeConfiguration(CONFIG_STORE_NAME, keys, reqMetadata).toIterable().iterator();
-		assertTrue(itr.hasNext());
-		SubscribeConfigurationResponse res = itr.next();
-		assertTrue(res.getItems().containsKey("configkey1"));
-		assertEquals("subscription_id", res.getSubscriptionId());
-		assertFalse(itr.hasNext());
-	}
-
-	@Test
-	public void subscribeConfigurationWithErrorTest() {
-		doAnswer((Answer<Void>) invocation -> {
-			StreamObserver<DaprProtos.SubscribeConfigurationResponse> observer =
-					(StreamObserver<DaprProtos.SubscribeConfigurationResponse>) invocation.getArguments()[1];
-			observer.onError(new RuntimeException());
-			observer.onCompleted();
-			return null;
-		}).when(daprStub).subscribeConfigurationAlpha1(any(DaprProtos.SubscribeConfigurationRequest.class), any());
-
-		assertThrowsDaprException(ExecutionException.class, () -> {
-			previewClient.subscribeConfiguration(CONFIG_STORE_NAME, "key").blockFirst();
-		});
-
-		assertThrows(IllegalArgumentException.class, () -> {
-			previewClient.subscribeConfiguration("", "key").blockFirst();
-		});
-	}
-
-	@Test
-	public void unsubscribeConfigurationTest() {
-		DaprProtos.UnsubscribeConfigurationResponse responseEnvelope = DaprProtos.UnsubscribeConfigurationResponse.newBuilder()
-				.setOk(true)
-				.setMessage("unsubscribed_message")
-				.build();
-
-		doAnswer((Answer<Void>) invocation -> {
-			StreamObserver<DaprProtos.UnsubscribeConfigurationResponse> observer =
-					(StreamObserver<DaprProtos.UnsubscribeConfigurationResponse>) invocation.getArguments()[1];
-			observer.onNext(responseEnvelope);
-			observer.onCompleted();
-			return null;
-		}).when(daprStub).unsubscribeConfigurationAlpha1(any(DaprProtos.UnsubscribeConfigurationRequest.class), any());
-
-		UnsubscribeConfigurationResponse
-				response = previewClient.unsubscribeConfiguration("subscription_id", CONFIG_STORE_NAME).block();
-		assertTrue(response.getIsUnsubscribed());
-		assertEquals("unsubscribed_message", response.getMessage());
-	}
-
-	@Test
-	public void unsubscribeConfigurationTestWithError() {
-		doAnswer((Answer<Void>) invocation -> {
-			StreamObserver<DaprProtos.UnsubscribeConfigurationResponse> observer =
-					(StreamObserver<DaprProtos.UnsubscribeConfigurationResponse>) invocation.getArguments()[1];
-			observer.onError(new RuntimeException());
-			observer.onCompleted();
-			return null;
-		}).when(daprStub).unsubscribeConfigurationAlpha1(any(DaprProtos.UnsubscribeConfigurationRequest.class), any());
-
-		assertThrowsDaprException(ExecutionException.class, () -> {
-			previewClient.unsubscribeConfiguration("subscription_id", CONFIG_STORE_NAME).block();
-		});
-
-		assertThrows(IllegalArgumentException.class, () -> {
-			previewClient.unsubscribeConfiguration("", CONFIG_STORE_NAME).block();
-		});
-
-		UnsubscribeConfigurationRequest req = new UnsubscribeConfigurationRequest("subscription_id", "");
-		assertThrows(IllegalArgumentException.class, () -> {
-			previewClient.unsubscribeConfiguration(req).block();
-		});
-	}
-
-	private DaprProtos.GetConfigurationResponse getSingleMockResponse() {
-		Map<String, String> metadata = new HashMap<>();
-		metadata.put("meta1", "value1");
-		Map<String, CommonProtos.ConfigurationItem> configs = new HashMap<>();
-		configs.put("configkey1", CommonProtos.ConfigurationItem.newBuilder()
-		.setValue("configvalue1")
-		.setVersion("1")
-		.putAllMetadata(metadata)
-		.build());
-		DaprProtos.GetConfigurationResponse responseEnvelope = DaprProtos.GetConfigurationResponse.newBuilder()
-				.putAllItems(configs)
-				.build();
-		return responseEnvelope;
-	}
-
-	private DaprProtos.GetConfigurationResponse getMultipleMockResponse() {
-		Map<String, String> metadata = new HashMap<>();
-		metadata.put("meta1", "value1");
-		Map<String, CommonProtos.ConfigurationItem> configs = new HashMap<>();
-		configs.put("configkey1", CommonProtos.ConfigurationItem.newBuilder()
-		.setValue("configvalue1")
-		.setVersion("1")
-		.putAllMetadata(metadata)
-		.build());
-		configs.put("configkey2", CommonProtos.ConfigurationItem.newBuilder()
-		.setValue("configvalue2")
-		.setVersion("1")
-		.putAllMetadata(metadata)
-		.build());
-		DaprProtos.GetConfigurationResponse responseEnvelope = DaprProtos.GetConfigurationResponse.newBuilder()
-				.putAllItems(configs)
-				.build();
-		return responseEnvelope;
 	}
 
 	@Test
